Safari-609.1.11.1/LayoutTests/inspector/controller/runtime-controller-import-expected.txt

CONSOLE MESSAGE: line 9: TypeError: Module specifier does not start with "/", "./", or "../".
CONSOLE MESSAGE: line 9: TypeError: Importing a module script failed.
CONSOLE MESSAGE: line 6: Cocoa is Sweet.
CONSOLE MESSAGE: line 7: %o
CONSOLE MESSAGE: line 6: Cocoa is Sweet. Cappuccino is Awesome.
CONSOLE MESSAGE: line 7: %o
Tests for RuntimeManager operations.


== Running test suite: RuntimeManager
-- Running test case: RuntimeManager.prototype.evaluateInInspectedWindow.ImportFromConsole
Source: await import('')
PASS: Transformed. Should log the value or an exception.
Source: await import('./not-found.js')
PASS: Transformed. Should log the value or an exception.
Source: await import('./resources/cocoa.js').then((cocoa) => { console.log(`${cocoa.name} is ${cocoa.taste}.`); })
PASS: Transformed. Should log the value or an exception.
Source: await import('./resources/drink.js').then((ns) => { console.log(`${ns.Cocoa.name} is ${ns.Cocoa.taste}. ${ns.Cappuccino.name} is ${ns.Cappuccino.taste}.`); })
PASS: Transformed. Should log the value or an exception.
